jdbc profiles

+++ src/main/java/org/saleslist/repository/jdbc/JdbcProductRepository.java
@@ -1,8 +1,9 @@
 package org.saleslist.repository.jdbc;
 
+import org.saleslist.Profiles;
 import org.saleslist.model.Product;
 import org.saleslist.repository.ProductRepository;
-import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.annotation.Profile;
 import org.springframework.dao.support.DataAccessUtils;
 import org.springframework.jdbc.core.BeanPropertyRowMapper;
 import org.springframework.jdbc.core.JdbcTemplate;
@@ -12,20 +13,19 @@ import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
 
+import java.sql.Timestamp;
 import java.time.LocalDateTime;
 import java.util.List;
 
 import static org.saleslist.web.SecurityUtil.ADMIN_ID;
 
-@Repository
-public class JdbcProductRepository implements ProductRepository {
+public abstract class JdbcProductRepository<T> implements ProductRepository {
 
     private final JdbcTemplate jdbcTemplate;
     private final NamedParameterJdbcTemplate namedParameterJdbcTemplate;
     private final SimpleJdbcInsert simpleJdbcInsert;
     private static final RowMapper<Product> ROW_MAPPER = BeanPropertyRowMapper.newInstance(Product.class);
 
-    @Autowired
     public JdbcProductRepository(JdbcTemplate jdbcTemplate, NamedParameterJdbcTemplate namedParameterJdbcTemplate) {
         this.jdbcTemplate = jdbcTemplate;
         this.namedParameterJdbcTemplate = namedParameterJdbcTemplate;
@@ -34,11 +34,40 @@ public class JdbcProductRepository implements ProductRepository {
                 .usingGeneratedKeyColumns("id");
     }
 
+    protected abstract T toDbDateTime(LocalDateTime ldt);
+
+    @Repository
+    @Profile(Profiles.POSTGRES_DB)
+    public static class Java8JdbcProductRepository extends JdbcProductRepository<LocalDateTime> {
+
+        public Java8JdbcProductRepository(JdbcTemplate jdbcTemplate, NamedParameterJdbcTemplate namedParameterJdbcTemplate) {
+            super(jdbcTemplate, namedParameterJdbcTemplate);
+        }
+
+        @Override
+        protected LocalDateTime toDbDateTime(LocalDateTime ldt) {
+            return ldt;
+        }
+    }
+
+    @Repository
+    @Profile(Profiles.HSQL_DB)
+    public static class TimestampJdbcProductRepository extends JdbcProductRepository<Timestamp> {
+        public TimestampJdbcProductRepository(JdbcTemplate jdbcTemplate, NamedParameterJdbcTemplate namedParameterJdbcTemplate) {
+            super(jdbcTemplate, namedParameterJdbcTemplate);
+        }
+
+        @Override
+        protected Timestamp toDbDateTime(LocalDateTime ldt) {
+            return Timestamp.valueOf(ldt);
+        }
+    }
+
     @Override
     public Product save(Product product, int userId) {
         MapSqlParameterSource map = new MapSqlParameterSource()
                 .addValue("id", product.getId())
-                .addValue("date_time", product.getDateTime())
+                .addValue("date_time", toDbDateTime(product.getDateTime()))
                 .addValue("title", product.getTitle())
                 .addValue("market_place", product.getMarketPlace().toString())
                 .addValue("delivery_service", product.getDeliveryService().toString())
@@ -58,21 +87,34 @@ public class JdbcProductRepository implements ProductRepository {
         } else {
             if (namedParameterJdbcTemplate.update(
                     "UPDATE products " +
-                    "SET date_time=:date_time, title=:title, market_place=:market_place, delivery_service=:delivery_service, payment_method=:payment_method, order_status=:order_status, sold_at_price=:sold_at_price, spent=:spent, payout_percentage=:payout_percentage, payout_currency=:payout_currency, profit=:profit, notes=:notes " +
-                    "WHERE id=:id AND user_id=:user_id", map) == 0) {
+                            "SET date_time=:date_time, title=:title, market_place=:market_place, delivery_service=:delivery_service, payment_method=:payment_method, order_status=:order_status, sold_at_price=:sold_at_price, spent=:spent, payout_percentage=:payout_percentage, payout_currency=:payout_currency, profit=:profit, notes=:notes " +
+                            "WHERE id=:id AND user_id=:user_id", map) == 0) {
                 return null;
             }
         }
 
         // for payout
+//        if (product.getPayoutPercentage() > 0) {
+//            jdbcTemplate.update("INSERT INTO payouts(user_id, product_id, date_time, amount, notes) " +
+//                    "VALUES (?,?,?,?,?) " +
+//                    "ON CONFLICT (product_id)" +
+//                    "DO UPDATE SET user_id=EXCLUDED.user_id, date_time=EXCLUDED.date_time, amount=EXCLUDED.amount, notes=EXCLUDED.notes",
+//                    userId, product.getId(), toDbDateTime(product.getDateTime()), product.getPayoutCurrency(), product.getTitle());
+//        } else {
+//            jdbcTemplate.update("DELETE FROM payouts WHERE product_id=?", product.getId());
+//        }
+
         if (product.getPayoutPercentage() > 0) {
+//            Payout payout = new Payout(product.getDateTime(), product.getPayoutCurrency(), product.getTitle());
             jdbcTemplate.update("INSERT INTO payouts(user_id, product_id, date_time, amount, notes) " +
-                    "VALUES (?,?,?,?,?) " +
-                    "ON CONFLICT (product_id)" +
-                    "DO UPDATE SET user_id=EXCLUDED.user_id, date_time=EXCLUDED.date_time, amount=EXCLUDED.amount, notes=EXCLUDED.notes", userId, product.getId(), product.getDateTime(), product.getPayoutCurrency(), product.getTitle());
-        } else {
+                            "VALUES (?,?,?,?,?) " +
+                            "ON CONFLICT (product_id)" +
+                            "DO UPDATE SET user_id=EXCLUDED.user_id, date_time=EXCLUDED.date_time, amount=EXCLUDED.amount, notes=EXCLUDED.notes",
+                    userId, product.getId(), toDbDateTime(product.getDateTime()), product.getPayoutCurrency(), product.getTitle());
+        } else if (get(product.getId(), userId).getPayoutPercentage() > 0 && product.getPayoutPercentage() == 0) {
             jdbcTemplate.update("DELETE FROM payouts WHERE product_id=?", product.getId());
         }
+
         return product;
     }
 
@@ -103,10 +145,12 @@ public class JdbcProductRepository implements ProductRepository {
     public List<Product> getBetween(LocalDateTime startDateTime, LocalDateTime endDateTime, int userId) {
         if (userId == ADMIN_ID) {
             return jdbcTemplate.query(
-                    "SELECT * FROM products WHERE date_time>=? AND date_time<? ORDER BY date_time DESC", ROW_MAPPER, startDateTime, endDateTime);
+                    "SELECT * FROM products WHERE date_time>=? AND date_time<? ORDER BY date_time DESC", ROW_MAPPER,
+                    toDbDateTime(startDateTime), toDbDateTime(endDateTime));
         }
         return jdbcTemplate.query(
-                "SELECT * FROM products WHERE user_id=? AND date_time>=? AND date_time<? ORDER BY date_time DESC", ROW_MAPPER, userId, startDateTime, endDateTime);
+                "SELECT * FROM products WHERE user_id=? AND date_time>=? AND date_time<? ORDER BY date_time DESC", ROW_MAPPER,
+                userId, startDateTime, endDateTime);
     }
 
     public List<String> getOwnersNames() {
